Consolidate various *Options classes into db_options
