+++ src/db.rs
@@ -16,10 +16,10 @@
 use crate::{
     ffi, ffi_util::opt_bytes_to_ptr, ColumnFamily, ColumnFamilyDescriptor, DBIterator,
     DBPinnableSlice, DBRawIterator, DBWALIterator, Direction, Error, FlushOptions, IteratorMode,
-    Options, Snapshot, WriteBatch, WriteOptions, DB,
+    Options, ReadOptions, Snapshot, WriteBatch, WriteOptions, DB,
 };
 
-use libc::{self, c_char, c_int, c_uchar, c_void, size_t};
+use libc::{self, c_char, c_int, c_void, size_t};
 use std::collections::BTreeMap;
 use std::ffi::{CStr, CString};
 use std::fmt;
@@ -37,36 +37,6 @@ unsafe impl Send for DB {}
 // Sync is similarly safe for many types because they do not expose interior mutability, and their
 // use within the rocksdb library is generally behind a const reference
 unsafe impl Sync for DB {}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub enum DBCompressionType {
-    None = ffi::rocksdb_no_compression as isize,
-    Snappy = ffi::rocksdb_snappy_compression as isize,
-    Zlib = ffi::rocksdb_zlib_compression as isize,
-    Bz2 = ffi::rocksdb_bz2_compression as isize,
-    Lz4 = ffi::rocksdb_lz4_compression as isize,
-    Lz4hc = ffi::rocksdb_lz4hc_compression as isize,
-    Zstd = ffi::rocksdb_zstd_compression as isize,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub enum DBCompactionStyle {
-    Level = ffi::rocksdb_level_compaction as isize,
-    Universal = ffi::rocksdb_universal_compaction as isize,
-    Fifo = ffi::rocksdb_fifo_compaction as isize,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub enum DBRecoveryMode {
-    TolerateCorruptedTailRecords = ffi::rocksdb_tolerate_corrupted_tail_records_recovery as isize,
-    AbsoluteConsistency = ffi::rocksdb_absolute_consistency_recovery as isize,
-    PointInTime = ffi::rocksdb_point_in_time_recovery as isize,
-    SkipAnyCorruptedRecord = ffi::rocksdb_skip_any_corrupted_records_recovery as isize,
-}
-
-pub struct ReadOptions {
-    pub(crate) inner: *mut ffi::rocksdb_readoptions_t,
-}
 
 impl DB {
     /// Open a database with default options.
@@ -940,105 +910,6 @@ impl fmt::Debug for DB {
         write!(f, "RocksDB {{ path: {:?} }}", self.path())
     }
 }
-
-impl Drop for ReadOptions {
-    fn drop(&mut self) {
-        unsafe { ffi::rocksdb_readoptions_destroy(self.inner) }
-    }
-}
-
-impl ReadOptions {
-    // TODO add snapshot setting here
-    // TODO add snapshot wrapper structs with proper destructors;
-    // that struct needs an "iterator" impl too.
-    #[allow(dead_code)]
-    fn fill_cache(&mut self, v: bool) {
-        unsafe {
-            ffi::rocksdb_readoptions_set_fill_cache(self.inner, v as c_uchar);
-        }
-    }
-
-    pub(crate) fn set_snapshot(&mut self, snapshot: &Snapshot) {
-        unsafe {
-            ffi::rocksdb_readoptions_set_snapshot(self.inner, snapshot.inner);
-        }
-    }
-
-    /// Set the upper bound for an iterator.
-    /// The upper bound itself is not included on the iteration result.
-    ///
-    /// # Safety
-    ///
-    /// This function will store a clone of key and will give a raw pointer of it to the
-    /// underlying C++ API, therefore, when given to any other [`DB`] method you must ensure
-    /// that this [`ReadOptions`] value does not leave the scope too early (e.g. `DB::iterator_cf_opt`).
-    pub unsafe fn set_iterate_upper_bound<K: AsRef<[u8]>>(&mut self, key: K) {
-        let key = key.as_ref();
-        ffi::rocksdb_readoptions_set_iterate_upper_bound(
-            self.inner,
-            key.as_ptr() as *const c_char,
-            key.len() as size_t,
-        );
-    }
-
-    pub fn set_prefix_same_as_start(&mut self, v: bool) {
-        unsafe { ffi::rocksdb_readoptions_set_prefix_same_as_start(self.inner, v as c_uchar) }
-    }
-
-    pub fn set_total_order_seek(&mut self, v: bool) {
-        unsafe { ffi::rocksdb_readoptions_set_total_order_seek(self.inner, v as c_uchar) }
-    }
-
-    /// If true, all data read from underlying storage will be
-    /// verified against corresponding checksums.
-    ///
-    /// Default: true
-    pub fn set_verify_checksums(&mut self, v: bool) {
-        unsafe {
-            ffi::rocksdb_readoptions_set_verify_checksums(self.inner, v as c_uchar);
-        }
-    }
-
-    /// If non-zero, an iterator will create a new table reader which
-    /// performs reads of the given size. Using a large size (> 2MB) can
-    /// improve the performance of forward iteration on spinning disks.
-    /// Default: 0
-    ///
-    /// ```
-    /// use rocksdb::{ReadOptions};
-    ///
-    /// let mut opts = ReadOptions::default();
-    /// opts.set_readahead_size(4_194_304); // 4mb
-    /// ```
-    pub fn set_readahead_size(&mut self, v: usize) {
-        unsafe {
-            ffi::rocksdb_readoptions_set_readahead_size(self.inner, v as size_t);
-        }
-    }
-
-    /// If true, create a tailing iterator. Note that tailing iterators
-    /// only support moving in the forward direction. Iterating in reverse
-    /// or seek_to_last are not supported.
-    pub fn set_tailing(&mut self, v: bool) {
-        unsafe {
-            ffi::rocksdb_readoptions_set_tailing(self.inner, v as c_uchar);
-        }
-    }
-}
-
-impl Default for ReadOptions {
-    fn default() -> ReadOptions {
-        unsafe {
-            ReadOptions {
-                inner: ffi::rocksdb_readoptions_create(),
-            }
-        }
-    }
-}
-
-unsafe impl Send for ReadOptions {}
-
-unsafe impl Sync for ReadOptions {}
 
 fn to_cpath<P: AsRef<Path>>(path: P) -> Result<CString, Error> {
     match CString::new(path.as_ref().to_string_lossy().as_bytes()) {
